Extract RC caching to a different class
