+++ malicious-site-protection/malicious-site-protection-impl/src/main/kotlin/com/duckduckgo/malicioussiteprotection/impl/RealMaliciousSiteProtection.kt
@@ -17,56 +17,17 @@
 package com.duckduckgo.malicioussiteprotection.impl
 
 import android.net.Uri
-import com.duckduckgo.app.di.AppCoroutineScope
-import com.duckduckgo.app.di.IsMainProcess
-import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.malicioussiteprotection.api.MaliciousSiteProtection
 import com.duckduckgo.malicioussiteprotection.api.MaliciousSiteProtection.IsMaliciousResult
-import com.duckduckgo.privacy.config.api.PrivacyConfigCallbackPlugin
 import com.squareup.anvil.annotations.ContributesBinding
-import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-import org.json.JSONObject
 import timber.log.Timber
 
 @ContributesBinding(AppScope::class, MaliciousSiteProtection::class)
-@ContributesMultibinding(AppScope::class, PrivacyConfigCallbackPlugin::class)
 class RealMaliciousSiteProtection @Inject constructor(
-    private val dispatchers: DispatcherProvider,
-    private val maliciousSiteProtectionFeature: MaliciousSiteProtectionFeature,
-    @IsMainProcess private val isMainProcess: Boolean,
-    @AppCoroutineScope private val appCoroutineScope: CoroutineScope,
-) : MaliciousSiteProtection, PrivacyConfigCallbackPlugin {
-
-    private var isFeatureEnabled = false
-
-    private var hashPrefixUpdateFrequency = 20L
-    private var filterSetUpdateFrequency = 720L
-
-    init {
-        if (isMainProcess) {
-            loadToMemory()
-        }
-    }
-
-    override fun onPrivacyConfigDownloaded() {
-        loadToMemory()
-    }
-
-    private fun loadToMemory() {
-        appCoroutineScope.launch(dispatchers.io()) {
-            isFeatureEnabled = maliciousSiteProtectionFeature.self().isEnabled()
-            maliciousSiteProtectionFeature.self().getSettings()?.let {
-                JSONObject(it).let { settings ->
-                    hashPrefixUpdateFrequency = settings.getLong("hashPrefixUpdateFrequency")
-                    filterSetUpdateFrequency = settings.getLong("filterSetUpdateFrequency")
-                }
-            }
-        }
-    }
+    maliciousSiteProtectionRCFeature: MaliciousSiteProtectionRCFeature,
+) : MaliciousSiteProtection {
 
     override suspend fun isMalicious(url: Uri, confirmationCallback: (isMalicious: Boolean) -> Unit): IsMaliciousResult {
         Timber.tag("MaliciousSiteProtection").d("isMalicious $url")
